Refactored Controller Classes
